refactored safenvml to simplify new api addition process.
1) use one api registry source for prod and tests
2) place safedevice interface and impl in separate files
3) add Readme.md to describe how a new API should be added
